Make it so only one tower can be dragged by a given input
This does give an advantage to touch-screen players, but I think
that's OK.

One thing: it seems like mouse+touch doesn't quite work, but this seems
to be because Windows moves the mouse to where a touch is, so the mouse
and touched ships follow the same point.

There may be a way to tell Windows not to do this, but I think it's
fine.