Try some different background gradient ideas
This one seems OK, although that might just because I'm working in
low brightness ATM.

Possible TODO: Make the planet rotate as well, or something?

Also, on that note: A real TODO is making lives actually be lost
when viruses get by, as well as a GAME OVER screen or something.